Refactoring of provenance.py, created submodule  with RO class definition
